fix(test): compile and run 32 tests that no module ever declared
Two sibling *_tests.rs files were left behind when their code moved, and
nothing declared either one. Cargo globs `tests/*.rs` but not inside `src/`,
so an undeclared unit-test file is silently never compiled. Both read as
coverage and verified nothing.

tools/schema_tests.rs (15 tests) — `schema.rs` became a re-export shim over
`tinyagents_harness::tool`, so `use super::*` no longer supplies the `json!`
macro the way it did when the implementation lived here. Wired via
`#[path]` and given the one missing import.

integrations/composio/bus_tests.rs (17 tests) — cannot be wired where it sat:
`integrations/composio/bus.rs` is a `pub use ...::bus::*` shim, and a glob
re-export cannot carry the private items these assert on
(`TRIAGE_DISABLED_ENV`, `triage_disabled`). Relocated to
memory/sync/composio/bus_tests.rs, beside the module that defines them, which
was already wired and held one unrelated test.

All 32 pass unmodified — the code had been relocated, not rewritten, so the
contracts they pin still hold. No assertion was changed, weakened or skipped.

Proven non-vacuous by fault injection rather than by their passing: dropping
"yes"/"YES" from the triage truthy list fails `triage_disabled_flag_parser`
on "expected 'yes' to disable triage", and removing "minLength" from
GEMINI_UNSUPPORTED_KEYWORDS fails three schema tests on their own assertions.
Both faults restored; 33 pass.
